+++ src/writer.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"bytes"
+	"errors"
 	"fmt"
 	"io/fs"
 	"io/ioutil"
@@ -15,19 +16,30 @@ import (
 const verbose = false
 
 type writer struct {
-	topFolder string
-	schema    Schema
-	module    string
+	topFolder, entityFolder         string
+	connectionFolder, supportFolder string
+	schema                          Schema
+	commandLine, module             string
+	connectionStringEnvArg          string
+	goFilesWritten                  []string
 }
 
 func NewWriter(
 	folder string,
 	module string,
+	commandLine string,
+	connectionStringEnvArg string,
 	schema Schema) writer {
 	w := writer{
-		topFolder: path.Clean(folder),
-		module:    module,
-		schema:    schema,
+		topFolder:              path.Clean(folder),
+		entityFolder:           path.Join(folder, "entities"),
+		connectionFolder:       path.Join(folder, "connection"),
+		supportFolder:          path.Join(folder, "support"),
+		module:                 module,
+		commandLine:            commandLine,
+		connectionStringEnvArg: connectionStringEnvArg,
+		schema:                 schema,
+		goFilesWritten:         []string{},
 	}
 	return w
 }
@@ -37,9 +49,15 @@ func (w writer) WriteStuff() {
 	w.createOutputFolders()
 	w.createDumpFile()
 
+	w.createSupportFile()
+	w.createEntities()
+	w.createConnection()
+
 	w.initModules()
 	w.fetchModules()
 	w.tidyModules()
+
+	w.applyFormatting()
 }
 
 func (w *writer) clearOutputFolder() {
@@ -50,12 +68,44 @@ func (w *writer) clearOutputFolder() {
 func (w *writer) createOutputFolders() {
 	fmt.Println("Ensuring destination folders exist")
 	check(os.MkdirAll(w.topFolder, 0755))
+	check(os.MkdirAll(w.supportFolder, 0755))
+	check(os.MkdirAll(w.entityFolder, 0755))
+	check(os.MkdirAll(w.connectionFolder, 0755))
 }
 
 func (w *writer) createDumpFile() {
 	fmt.Println("Writing JSON dump file")
 	filename := path.Join(w.topFolder, "dump.json")
 	check(ioutil.WriteFile(filename, w.schema.ToJSON(), fs.ModePerm))
+}
+
+func (w *writer) createSupportFile() {
+	fmt.Println("Creating support file")
+	filename := path.Join(w.supportFolder, "support.go")
+	w.writeGoFile(filename, "support", nil)
+}
+
+func (w *writer) createEntities() {
+	fmt.Println("Creating entities")
+	for _, table := range w.schema.Tables {
+		hasPrimary := false
+		for _, c := range table.Columns {
+			if c.IsPrimaryKey {
+				hasPrimary = true
+			}
+		}
+		if !hasPrimary {
+			check(errors.New(fmt.Sprintf("%s.%s has no primary key", w.schema.SchemaName, table.TableName)))
+		}
+		filename := path.Join(w.entityFolder, table.SlugName+".go")
+		w.writeGoFile(filename, "entities", table)
+	}
+}
+
+func (w *writer) createConnection() {
+	fmt.Println("Creating connection")
+	filename := path.Join(w.connectionFolder, "connection.go")
+	w.writeGoFile(filename, "connection", nil)
 }
 
 func (w *writer) initModules() {
@@ -98,4 +148,21 @@ func (w *writer) runCmd(display bool, folder string, command string, args ...str
 	if verbose && out.Len() > 0 {
 		fmt.Println(out.String())
 	}
+}
+
+func (w *writer) applyFormatting() {
+	fmt.Println("Formatting generated Go source")
+	for _, filename := range w.goFilesWritten {
+		w.runCmd(true, "", "gofmt", "-w", "-s", filename)
+	}
+}
+
+func (w *writer) writeGoFile(filename string, templateName string, data interface{}) {
+	w.writeFile(filename, templateName, data)
+	w.goFilesWritten = append(w.goFilesWritten, filename)
+}
+
+func (w *writer) writeFile(filename string, templateName string, data interface{}) {
+	b := w.getTemplatedData(data, templateName)
+	check(ioutil.WriteFile(filename, b, fs.ModePerm))
 }